Seed grades from pre-filled points, fix tab routing, auto AI-suggest
